+++ text2sparql_client/commands/ask.py
@@ -1,6 +1,7 @@
 """query command"""
 
 import json
+import sys
 from io import TextIOWrapper
 from pathlib import Path
 
@@ -13,6 +14,13 @@ from pydantic import ValidationError
 from text2sparql_client.database import Database
 from text2sparql_client.models.questions_file import QuestionsFile
 from text2sparql_client.request import text2sparql
+
+
+def check_output_file(file: str) -> None:
+    """Check if output file already exists."""
+    if Path(file).exists():
+        logger.error(f"Output file {file} already exists.")
+        sys.exit(1)
 
 
 @click.command(name="ask")
@@ -28,15 +36,24 @@ from text2sparql_client.request import text2sparql
 @click.option(
     "--timeout",
     type=int,
-    default=120,
+    default=600,
     show_default=True,
     help="Timeout in seconds.",
+)
+@click.option(
+    "--output",
+    "-o",
+    type=click.Path(dir_okay=False, writable=True, file_okay=True, allow_dash=True),
+    default="-",
+    show_default=True,
+    help="Save JSON output to this file.",
 )
 def ask_command(
     questions_file: TextIOWrapper,
     url: str,
     answers_db: str,
     timeout: int,
+    output: str,
 ) -> None:
     """Query a TEXT2SPARQL endpoint
 
@@ -46,6 +63,7 @@ def ask_command(
     database = Database(file=Path(answers_db))
     file_model = QuestionsFile.model_validate(yaml.safe_load(questions_file))
     logger.info(f"Asking questions about dataset {file_model.dataset.id} on endpoint {url}.")
+    check_output_file(file=output)
     responses = []
     for question_section in file_model.questions:
         for language, question in question_section.question.items():
@@ -70,4 +88,7 @@ def ask_command(
             except ValidationError as error:
                 logger.debug(str(error))
                 logger.error("validation error")
-    click.secho(json.dumps(responses, indent=2))
+    check_output_file(file=output)
+    logger.info(f"Writing {len(responses)} responses to {output if output != '-' else 'stdout'}.")
+    with click.open_file(filename=output, mode="w", encoding="UTF-8") as file:
+        json.dump(responses, file, indent=2)