+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [1.2.0] 2025-04-22
+
+### Added
+
+- ask command
+  - `--output` / `-o` with default value `-` for stdout
+- proper test suite
+
 ## [1.1.0] 2025-04-18
 
 ### Added